+++ packages/ddp/stream_server.js
@@ -1,5 +1,30 @@
 var url = Npm.require('url');
 
+// By default, we use the permessage-deflate extension with default
+// configuration. If $SERVER_WEBSOCKET_COMPRESSION is set, then it must be valid
+// JSON. If it represents a falsey value, then we do not use permessage-deflate
+// at all; otherwise, the JSON value is used as an argument to deflate's
+// configure method; see
+// https://github.com/faye/permessage-deflate-node/blob/master/README.md
+//
+// (We do this in an _.once instead of at startup, because we don't want to
+// crash the tool during isopacket load if your JSON doesn't parse. This is only
+// a problem because the tool has to load the DDP server code just in order to
+// be a DDP client; see https://github.com/meteor/meteor/issues/3452 .)
+var websocketExtensions = _.once(function () {
+  var extensions = [];
+
+  var websocketCompressionConfig = process.env.SERVER_WEBSOCKET_COMPRESSION
+        ? JSON.parse(process.env.SERVER_WEBSOCKET_COMPRESSION) : {};
+  if (websocketCompressionConfig) {
+    extensions.push(Npm.require('permessage-deflate').configure(
+      websocketCompressionConfig
+    ));
+  }
+
+  return extensions;
+});
+
 var pathPrefix = __meteor_runtime_config__.ROOT_URL_PATH_PREFIX ||  "";
 
 StreamServer = function () {
@@ -41,8 +66,13 @@ StreamServer = function () {
   // from ever working, set $DISABLE_WEBSOCKETS and SockJS clients (ie,
   // browsers) will not waste time attempting to use them.
   // (Your server will still have a /websocket endpoint.)
-  if (process.env.DISABLE_WEBSOCKETS)
+  if (process.env.DISABLE_WEBSOCKETS) {
     serverOptions.websocket = false;
+  } else {
+    serverOptions.faye_server_options = {
+      extensions: websocketExtensions()
+    };
+  }
 
   self.server = sockjs.createServer(serverOptions);
   if (!Package.webapp) {